Восстановление data.js
