+++ js/api.js
@@ -1,10 +1,7 @@
 import {
   createPointsOfMap
 } from './map.js';
-import {
-  closeSuccessModalElement,
-  closeErrorModalElement
-} from './utilities.js';
+
 const getData = () => {
   fetch('https://24.javascript.pages.academy/keksobooking/data')
     .then((response) => response.json())
@@ -27,16 +24,14 @@ const sendData = (onSuccess, onError) => {
     }).then((response) => {
       if (response.ok) {
         onSuccess();
-        closeSuccessModalElement();
       } else {
         onError();
-        closeErrorModalElement();
       }
     }).catch(() => {
       onError();
-      closeErrorModalElement();
     });
   });
+
 };
 
 export {
